+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Both client libraries are pre-1.0, and they have separate versioning.
 
 ## Unreleased
 
-No unreleased changes.
+- `Sandbox.exec` in JS now correctly accepts a list of Secrets.
 
 ## modal-js/v0.3.18, modal-go/v0.0.18